+++ semcore/data-table/CHANGELOG.md
@@ -6,7 +6,7 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 
 ### Added
 
-- `rowsRendering` prop to `DataTable.Body` for low level control over rows rendering. Especially it allows to use custom virtual rendering.
+- `renderRows` prop to `DataTable.Body` for low level control over rows rendering. Especially it allows to use custom virtual rendering.
 
 ## [4.28.4] - 2024-03-11
 
@@ -18,7 +18,7 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 
 ### Changed
 
-- Don't focus hidden DataTable.Head.
+- Don't focus hidden DataTable. Head.
 
 ## [4.28.2] - 2024-03-07